change pl parseint to parsefloat

+++ src/components/gameDetails/manulMarkets/index.tsx
@@ -115,7 +115,7 @@ const ManualMarket = ({ title, data, detail }) => {
             <div className="d-flex flex-row justify-content-between w-100">
               <span
                 className={`${
-                  parseInt(
+                  parseFloat(
                     detail?.profitLossDataMatch?.[
                       profitLossDataForMatchConstants[data?.type]?.A +
                         "_" +
@@ -146,7 +146,7 @@ const ManualMarket = ({ title, data, detail }) => {
                         detail?.id
                     ] === "0"
                     ? ""
-                    : parseInt(
+                    : parseFloat(
                         detail?.profitLossDataMatch?.[
                           profitLossDataForMatchConstants[data?.type]?.A +
                             "_" +
@@ -356,7 +356,7 @@ const ManualMarket = ({ title, data, detail }) => {
             <div className="d-flex flex-row justify-content-between w-100">
               <span
                 className={`${
-                  parseInt(
+                  parseFloat(
                     detail?.profitLossDataMatch?.[
                       profitLossDataForMatchConstants[data?.type]?.B +
                         "_" +
@@ -387,7 +387,7 @@ const ManualMarket = ({ title, data, detail }) => {
                         detail?.id
                     ] === "0"
                     ? ""
-                    : parseInt(
+                    : parseFloat(
                         detail?.profitLossDataMatch?.[
                           profitLossDataForMatchConstants[data?.type]?.B +
                             "_" +
@@ -589,7 +589,7 @@ const ManualMarket = ({ title, data, detail }) => {
               <div className="d-flex flex-row justify-content-between w-100">
                 <span
                   className={`${
-                    parseInt(
+                    parseFloat(
                       detail?.profitLossDataMatch?.[
                         profitLossDataForMatchConstants[data?.type]?.C +
                           "_" +
@@ -618,7 +618,7 @@ const ManualMarket = ({ title, data, detail }) => {
                           detail?.id
                       ] === "0"
                       ? ""
-                      : parseInt(
+                      : parseFloat(
                           detail?.profitLossDataMatch?.[
                             profitLossDataForMatchConstants[data?.type]?.C +
                               "_" +
